+++ pyemma/msm/models/msm.py
@@ -69,12 +69,12 @@ class MSM(_Model):
             number, whitespace and unit. Permitted units are
             (* is an arbitrary string):
 
-            |  'fs',  'femtosecond*'
-            |  'ps',  'picosecond*'
-            |  'ns',  'nanosecond*'
-            |  'us',  'microsecond*'
-            |  'ms',  'millisecond*'
-            |  's',   'second*'
+            *  'fs',  'femtosecond*'
+            *  'ps',  'picosecond*'
+            *  'ns',  'nanosecond*'
+            *  'us',  'microsecond*'
+            *  'ms',  'millisecond*'
+            *  's',   'second*'
 
         neig : int or None
             The number of eigenvalues / eigenvectors to be kept. If set to None,
@@ -119,12 +119,12 @@ class MSM(_Model):
             physical time unit. Specify by a number, whitespace and unit.
             Permitted units are (* is an arbitrary string):
 
-            |  'fs',  'femtosecond*'
-            |  'ps',  'picosecond*'
-            |  'ns',  'nanosecond*'
-            |  'us',  'microsecond*'
-            |  'ms',  'millisecond*'
-            |  's',   'second*'
+            *  'fs',  'femtosecond*'
+            *  'ps',  'picosecond*'
+            *  'ns',  'nanosecond*'
+            *  'us',  'microsecond*'
+            *  'ms',  'millisecond*'
+            *  's',   'second*'
 
         neig : int or None
             The number of eigenvalues / eigenvectors to be kept. If set to
@@ -443,7 +443,7 @@ class MSM(_Model):
 
         """
         p0 = _types.ensure_ndarray(p0, ndim=1, size=self.nstates, kind='numeric')
-        assert _types.is_int(k) and k>=0, 'k must be a non-negative integer'
+        assert _types.is_int(k) and k >= 0, 'k must be a non-negative integer'
 
         if k == 0:  # simply return p0 normalized
             return p0 / p0.sum()